backend_services/account_serv/requirements.txt
asgiref==3.7.2
black==23.9.1
cfgv==3.4.0
click==8.1.7
dataclasses==0.6
dataclasses-json==0.6.1
Deprecated==1.2.14
distlib==0.3.7
Django==4.2.6
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.12.4
flake8==6.1.0
googleapis-common-protos==1.56.2
grpcio==1.59.0
identify==2.5.30
importlib-metadata==6.8.0
iniconfig==2.0.0
isort==5.12.0
marshmallow==3.20.1
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
opentelemetry-api==1.20.0
opentelemetry-exporter-jaeger==1.20.0
opentelemetry-exporter-jaeger-proto-grpc==1.20.0
opentelemetry-exporter-jaeger-thrift==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-dbapi==0.41b0
opentelemetry-instrumentation-django==0.41b0
opentelemetry-instrumentation-psycopg2==0.41b0
opentelemetry-instrumentation-redis==0.41b0
opentelemetry-instrumentation-wsgi==0.41b0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
pre-commit==3.4.0
protobuf==3.20.3
psycopg2-binary==2.9.9
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
pytest==7.4.2
pytest-django==4.5.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
structlog==23.1.0
thrift==0.16.0
typing-inspect==0.9.0
typing_extensions==4.8.0
virtualenv==20.24.5
wrapt==1.15.0
zipp==3.17.0
